backend/app/core/config.py
# backend/app/core/config.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv(dotenv_path=os.path.join(os.path.dirname(__file__), '../../.env'))

class Settings:
    DATABASE_URL: str = os.getenv("DATABASE_URL")

settings = Settings()
