backend/requirements.txt
# Core packages
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.31
python-dotenv==1.0.1
psycopg2-binary==2.9.3
alembic==1.7.7

# Additional packages
requests==2.31.0
spotipy==2.23.0
yt-dlp==2023.12.30
boto3==1.28.0  # AWS S3
websockets==12.0
jwt==1.3.1

# Dependency management
httpx==0.26.0
httpcore==1.0.2
certifi==2023.11.17
idna==3.6
sniffio==1.3.0
h11==0.14.0

# Utility packages
pydantic==2.5.3
gunicorn==20.1.0  # For production deployment
